requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
aiohappyeyeballs=2.4.0=pypi_0
aiohttp=3.10.6=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.7.0=pypi_0
async-timeout=4.0.3=pypi_0
attrs=24.2.0=pypi_0
bzip2=1.0.8=h80987f9_6
ca-certificates=2024.7.2=hca03da5_0
cachetools=5.5.0=pypi_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
eval-type-backport=0.2.0=pypi_0
filelock=3.16.1=pypi_0
frozenlist=1.4.1=pypi_0
google-ai-generativelanguage=0.6.10=pypi_0
google-api-core=2.20.0=pypi_0
google-api-python-client=2.147.0=pypi_0
google-auth=2.35.0=pypi_0
google-auth-httplib2=0.2.0=pypi_0
google-generativeai=0.8.2=pypi_0
googleapis-common-protos=1.65.0=pypi_0
grpcio=1.66.1=pypi_0
grpcio-status=1.66.1=pypi_0
httplib2=0.22.0=pypi_0
idna=3.10=pypi_0
libffi=3.4.4=hca03da5_1
lxml=5.3.0=pypi_0
markdown-it-py=3.0.0=pypi_0
mdurl=0.1.2=pypi_0
multidict=6.1.0=pypi_0
ncurses=6.4=h313beb8_0
numpy=2.1.1=pypi_0
openssl=3.0.15=h80987f9_0
pillow=10.4.0=pypi_0
pip=24.2=py310hca03da5_0
proto-plus=1.24.0=pypi_0
protobuf=5.28.2=pypi_0
pyarrow=17.0.0=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pydantic=2.9.2=pypi_0
pydantic-core=2.23.4=pypi_0
pygments=2.18.0=pypi_0
pyparsing=3.1.4=pypi_0
python=3.10.14=hb885b13_1
python-dotenv=1.0.1=pypi_0
python-pptx=1.0.2=pypi_0
readline=8.2=h1a28f6b_0
requests=2.32.3=pypi_0
rich=13.8.1=pypi_0
rsa=4.9=pypi_0
setuptools=75.1.0=py310hca03da5_0
shellingham=1.5.4=pypi_0
sqlite=3.45.3=h80987f9_0
tabulate=0.9.0=pypi_0
tk=8.6.14=h6ba3021_0
together=1.3.0=pypi_0
tqdm=4.66.5=pypi_0
typer=0.12.5=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024a=h04d1e81_0
uritemplate=4.1.1=pypi_0
urllib3=2.2.3=pypi_0
wheel=0.44.0=py310hca03da5_0
xlsxwriter=3.2.0=pypi_0
xz=5.4.6=h80987f9_1
yarl=1.12.1=pypi_0
zlib=1.2.13=h18a0788_1
